Public Linked Navigation
Ok. So I've linked the navigation based on the pages that were there.
If I couldn't find your page-- i made up a link.   So everyone check to
make sure your links are right please..  And please let me know if any
are missing.

I will be doing the admin side soon, so the same applies. But i will let
you know when the nav is done.. because I might not get to it tonight.