Fixed admin home page
Take a look to see if i got the description right.   This set up looked
a lot nicer than sticking a ton of hyperlinks in a list.
ALl you haveto do is click on the page title, and it will drop down to
show what is on that page, and has a button that is linked to the page
itself.